TEIID-3503: breaking the jboss eap dependencies from admin module into separate module called jboss-admin, rearranged the follow on dependency modules to depend also on jboss-admin module
