+++ jboss-admin/src/main/java/org/teiid/adminapi/jboss/MetadataMapper.java
@@ -19,7 +19,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
  * 02110-1301 USA.
  */
-package org.teiid.adminapi.impl;
+package org.teiid.adminapi.jboss;
 
 import org.jboss.dmr.ModelNode;
