+++ jboss-admin/src/test/java/org/teiid/adminapi/jboss/TestSessionMetadata.java
@@ -20,12 +20,14 @@
  * 02110-1301 USA.
  */
 
-package org.teiid.adminapi.impl;
+package org.teiid.adminapi.jboss;
 
 import static org.junit.Assert.assertEquals;
 
 import org.jboss.dmr.ModelNode;
 import org.junit.Test;
+import org.teiid.adminapi.impl.SessionMetadata;
+import org.teiid.adminapi.jboss.VDBMetadataMapper;
 
 @SuppressWarnings("nls")
 public class TestSessionMetadata {
