+++ jboss-admin/src/test/resources/vdb-describe.txt
