+++ runtime/pom.xml
@@ -112,6 +112,10 @@
     </dependency>
     <dependency>
         <groupId>org.jboss</groupId>
+        <artifactId>jboss-vfs</artifactId>
+    </dependency>    
+    <dependency>   
+        <groupId>org.jboss</groupId>
         <artifactId>staxmapper</artifactId>
     </dependency>    
     <dependency>
