+++ test-integration/common/src/test/java/org/teiid/arquillian/AdminUtil.java
@@ -28,8 +28,8 @@ import java.util.Properties;
 import org.teiid.adminapi.Admin;
 import org.teiid.adminapi.AdminException;
 import org.teiid.adminapi.VDB;
-import org.teiid.adminapi.AdminFactory.AdminImpl;
 import org.teiid.adminapi.VDB.Status;
+import org.teiid.adminapi.jboss.AdminFactory.AdminImpl;
 
 @SuppressWarnings("nls")
 public class AdminUtil {
